requirements.txt
fastapi
pydantic
uvicorn
openai
python-dotenv
